UPDATE: package passprint. path to save leys modified and check if pa…

+++ utils/passprintPackage/backend/keys/savePrivateAndPublicKeyInFiles.ts
@@ -1,7 +1,8 @@
-import { writeFile, chmod } from 'node:fs/promises';
+import { writeFile, mkdir, chmod } from 'node:fs/promises';
 import { join } from 'node:path';
 import { type KeyPair } from './generatePrivateAndPublicKey';
-import { config } from './configurations';
+import { config } from '../config/configurations';
+import { existsSync } from 'node:fs';
 
 /**
  * Sauvegarde la paire de clés (publique et privée) dans des fichiers.
@@ -27,21 +28,32 @@ export async function savePrivateAndPublicKeyInFiles(keys: KeyPair) {
       config.privateKeyFile.path,
       config.privateKeyFile.name,
     );
-    await writeFile(privateKeyPath, privateKey, 'utf8');
-    await chmod(privateKeyPath, 0o600); // Utiliser la notation octale 0o600
-    console.log(
-      `Clé privée sauvegardée et permissions définies sur 600 : ${privateKeyPath}`,
-    );
 
-    const publicKeyPath = join(
-      config.publicKeyFile.path,
-      config.publicKeyFile.name,
-    );
-    await writeFile(publicKeyPath, publicKey, 'utf8');
-    await chmod(publicKeyPath, 0o644);
-    console.log(
-      `Clé publique sauvegardée et permissions définies sur 644 : ${publicKeyPath}`,
-    );
+    const keysFilesAlreadyExist =
+      existsSync(privateKeyPath) ||
+      existsSync(join(config.publicKeyFile.path, config.publicKeyFile.name));
+    if (keysFilesAlreadyExist) {
+      console.log(
+        "Les fichiers de clés existent déjà. Abandon de la sauvegarde pour éviter l'écrasement.",
+      );
+    } else {
+      await mkdir(config.privateKeyFile.path, { recursive: true });
+      await writeFile(privateKeyPath, privateKey, 'utf8');
+      await chmod(privateKeyPath, 0o600); // Utiliser la notation octale 0o600
+      console.log(
+        `Clé privée sauvegardée et permissions définies sur 600 : ${privateKeyPath}`,
+      );
+
+      const publicKeyPath = join(
+        config.publicKeyFile.path,
+        config.publicKeyFile.name,
+      );
+      await writeFile(publicKeyPath, publicKey, 'utf8');
+      await chmod(publicKeyPath, 0o644);
+      console.log(
+        `Clé publique sauvegardée et permissions définies sur 644 : ${publicKeyPath}`,
+      );
+    }
   } catch (err) {
     console.error('Erreur lors de la sauvegarde des clés :', err);
     throw new Error(
